+++ lib/src/controller.dart
@@ -252,6 +252,9 @@ class IntroController {
               _buildBorder(context),
               _buildHighlight(context),
               _buildCard(context),
+              ...intro.topLayerBuilder == null ? [] : [
+                intro.topLayerBuilder!(context, this),
+              ],
             ],
           ),
         ),